switch to lib repository and add shop in me api
