+++ user/pkg/userproto/user.proto
@@ -15,6 +15,7 @@ message User {
   string role = 2;
   string email = 3;
   string google_id = 4;
+  optional int32 shop_id = 5;
 }
 
 message MeReq {}
